% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandTransData.R
\name{expandTransData}
\alias{expandTransData}
\title{Reshape Data to 'Wide' Data Format}
\usage{
expandTransData(transData, wgtname)
}
\arguments{
\item{transData}{dataframe containing the time series vector of transition counts.}

\item{wgtname}{weight.}
}
\value{
The output a dataframe of transition data in 'wide' format.
}
\description{
This function reshapes time series vector of transition counts into a 'Wide' Data Format. Each non-Zero weighted row is
 expanded to show one row per record.
}
\author{
Abdoulaye (Ab) N'Diaye
}
