% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlabToPOSIX.R
\name{matlabToPOSIX}
\alias{matlabToPOSIX}
\title{Convert a numeric  MATLAB datenum to R POSIXt time values}
\usage{
matlabToPOSIX(gTime, timez)
}
\arguments{
\item{gTime}{a MATLAB datenum value}

\item{timez}{time zone, default is "UTC"}
}
\value{
R POSIXt time values.
}
\description{
This function is used to convert a numeric  MATLAB datenum
}
\examples{

matlabToPOSIX(734139)

}
\author{
Abdoulaye (Ab) N'Diaye
}
