% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSIXTomatlab.R
\name{POSIXTomatlab}
\alias{POSIXTomatlab}
\title{Convert between MATLAB datenum and R POSIXt}
\usage{
POSIXTomatlab(gTime)
}
\arguments{
\item{gTime}{a POSIXct or POSIXlt date value}
}
\value{
MATLAB datenum value.
}
\description{
This function is used to convert between MATLAB datenum values and R POSIXt time values.
}
\examples{

POSIXTomatlab(as.POSIXlt(as.Date("2010-01-01")))

}
\author{
Abdoulaye (Ab) N'Diaye
}
