% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTransProb.R
\docType{package}
\name{predData_ann_SeverelyAdverse}
\alias{predData_ann_SeverelyAdverse}
\alias{predData_ann_SeverelyAdverse-package}
\title{predData_ann_SeverelyAdverse}
\description{
predData_ann_SeverelyAdverse
}
