% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTransProb.R
\docType{package}
\name{predData_lda_SeverelyAdverse}
\alias{predData_lda_SeverelyAdverse}
\alias{predData_lda_SeverelyAdverse-package}
\title{predData_lda_SeverelyAdverse}
\description{
predData_lda_SeverelyAdverse
}
