% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTransProb.R
\docType{package}
\name{predData_mnl_SeverelyAdverse}
\alias{predData_mnl_SeverelyAdverse}
\alias{predData_mnl_SeverelyAdverse-package}
\title{predData_mnl_SeverelyAdverse}
\description{
predData_mnl_SeverelyAdverse
}
