% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTransProb.R
\docType{package}
\name{predData_svm_SeverelyAdverse}
\alias{predData_svm_SeverelyAdverse}
\alias{predData_svm_SeverelyAdverse-package}
\title{predData_svm_SeverelyAdverse}
\description{
predData_svm_SeverelyAdverse
}
