\name{RTransProb-Internal}
\alias{RTransProb-Internal}
\alias{cm.matrix.counts}
\alias{cm.matrix}
\alias{cm.vector.counts}
\alias{getidTotCntCohort}
\alias{getidTotCntDuration} 
\alias{is.leapyear}
\alias{yearFractionLocal}
\alias{getRatingsSnapshotsLocal}
\alias{getSampleTotals}
\alias{getSampleTotalsDur}
\alias{getCategoriesLocal}
\alias{getTransitionProbability}
\alias{fixDate}

\title{Internal RTransProb Functions
}
\description{ 
          Internal RTransProb Functions 
}

\keyword{ internal }% use one of  RShowDoc("KEYWORDS")

