% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VecOfTransData.R
\name{VecOfTransData}
\alias{VecOfTransData}
\title{Vector of Transition Counts}
\usage{
VecOfTransData(lstCnt,ratingCat,startDate,endDate,snapshots)
}
\arguments{
\item{lstCnt}{quarterly transition count data.}

\item{ratingCat}{list containing the unique rating caetgories.}

\item{startDate}{start date of the estimation time window, in string or numeric format. The default start date is the earliest date in 'data'.}

\item{endDate}{end date of the estimation time window, in string or numeric format. The default end date is the latest date in 'data'. The end date cannot be a date before the start date.}

\item{snapshots}{Integer indicating the number of credit-rating snapshots per year to be considered for the estimation.
Valid values are 1, 4, 12, 54, and 356.
For example, 1 = one snapshot per year}
}
\value{
The output is a time series vector of transition counts .
}
\description{
This function reshapes transition matrices of counts into a 'wide format', time series of transition counts.
}
\author{
Abdoulaye (Ab) N'Diaye
}
